Scale TxnFields

+++ evm/src/cpu/kernel/constants/mod.rs
@@ -61,14 +61,15 @@ pub fn evm_constants() -> HashMap<String, U256> {
         c.insert(segment.var_name().into(), (segment as u64).into());
     }
     for txn_field in NormalizedTxnField::all() {
-        c.insert(txn_field.var_name().into(), (txn_field as u32).into());
+        // These offsets are already scaled by their respective segment.
+        c.insert(txn_field.var_name().into(), (txn_field as usize).into());
     }
     for txn_field in GlobalMetadata::all() {
-        // Those offsets are already scaled by their respective segment.
+        // These offsets are already scaled by their respective segment.
         c.insert(txn_field.var_name().into(), (txn_field as usize).into());
     }
     for txn_field in ContextMetadata::all() {
-        // Those offsets are already scaled by their respective segment.
+        // These offsets are already scaled by their respective segment.
         c.insert(txn_field.var_name().into(), (txn_field as usize).into());
     }
     for trie_type in PartialTrieType::all() {
